WIP MFA TOTP Support. Interactions are now more generic, amr actions and amr registrations.

+++ server/routes/user.ts
@@ -22,7 +22,7 @@ userRouter.use(checkLoggedIn)
 
 userRouter.get('/me',
   (req, res) => {
-    res.send(req.user)
+    res.send(req.loggedInUser)
   },
 )
 
@@ -31,7 +31,12 @@ userRouter.patch('/profile',
     name: nameValidation,
   }),
   async (req, res) => {
-    const user = req.user
+    const user = req.loggedInUser
+    if (!user) {
+      res.sendStatus(500)
+      return
+    }
+
     const profile = validatorData<UpdateProfile>(req)
 
     await db().table<User>(TABLES.USER).update(profile).where({ id: user.id })
@@ -45,7 +50,11 @@ userRouter.patch('/email',
     email: emailValidation,
   }),
   async (req, res) => {
-    const user = req.user
+    const user = req.loggedInUser
+    if (!user) {
+      res.sendStatus(500)
+      return
+    }
 
     const { email } = validatorData<UpdateEmail>(req)
 
@@ -69,7 +78,12 @@ userRouter.patch('/password',
     newPassword: newPasswordValidation,
   }),
   async (req, res) => {
-    const user = req.user
+    const user = req.loggedInUser
+    if (!user) {
+      res.sendStatus(500)
+      return
+    }
+
     const { oldPassword, newPassword } = validatorData<UpdatePassword>(req)
 
     if (user.hasPassword && (!oldPassword || !await checkPasswordHash(user.id, oldPassword))) {
@@ -83,7 +97,11 @@ userRouter.patch('/password',
 )
 
 userRouter.delete('/passkeys', async (req, res) => {
-  const user = req.user
+  const user = req.loggedInUser
+  if (!user) {
+    res.sendStatus(500)
+    return
+  }
 
   if (!user.hasPassword) {
     res.sendStatus(400)
@@ -100,7 +118,11 @@ userRouter.delete('/passkeys', async (req, res) => {
 })
 
 userRouter.delete('/password', async (req, res) => {
-  const user = req.user
+  const user = req.loggedInUser
+  if (!user) {
+    res.sendStatus(500)
+    return
+  }
 
   if (!user.hasPasskeys) {
     res.sendStatus(400)
@@ -118,9 +140,13 @@ userRouter.delete('/password', async (req, res) => {
 })
 
 userRouter.delete('/totp', async (req, res) => {
-  const user = req.user
+  const user = req.loggedInUser
+  if (!user) {
+    res.sendStatus(500)
+    return
+  }
 
-  if (!user.mfaEnabled) {
+  if (!user.hasTotp) {
     res.sendStatus(404)
     return
   }
@@ -130,7 +156,11 @@ userRouter.delete('/totp', async (req, res) => {
 })
 
 userRouter.delete('/user', async (req, res) => {
-  const user = req.user
+  const user = req.loggedInUser
+  if (!user) {
+    res.sendStatus(500)
+    return
+  }
 
   await db().table<User>(TABLES.USER).delete().where({ id: user.id })
   await db().table<OIDCPayload>(TABLES.OIDC_PAYLOADS).delete().where({ accountId: user.id })
